[11/08/2012] Version 0.0.901.234a Global Project revue : - new project version - new GUI interface - improvements : very easy to use, fast scan, fast results, more memory - and...

+++ trunk/RtCA_Frameworks/RtCA_Full/tests/test_log_files_evtx.c
@@ -197,7 +197,7 @@ unsigned int ReadRecord(char *buffer, DWORD size, STRING_TABLE *my_s_table, char
         {
           //id identifié ^^
           mid = (M_ID *)c;
-          snprintf(log_id,DEFAULT_TMP_SIZE,"%08lu",mid->id);
+          snprintf(log_id,DEFAULT_TMP_SIZE,"%08lu",(DWORD) mid->id);
           id = mid->id;
 
           //Type
